INTYGFV-15858: Divided factories
